+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changes
 =======
 
+Version 0.9.3 (UNRELEASED)
+--------------------------
+
+- Changes version of NGINX Docker image from 1.19 to 1.25.
+
 Version 0.9.2 (2023-12-06)
 --------------------------